requirements.txt
aiogram
python-dotenv
binance-futures-connector
cryptography